ast_analysis.py
import ast

class ASTAnalyzer(ast.NodeVisitor):
    def __init__(self):
        self.max_depth = 0
        self.current_depth = 0
        self.identifier_length_violation = False

    def visit_Name(self, node):
        if len(node.id) == 13:  # Check if identifier length is exactly 13
            self.identifier_length_violation = True
        self.generic_visit(node)

    def visit(self, node):
        """Overriding visit to handle depth calculation."""
        if isinstance(node, (ast.If, ast.For, ast.While, ast.Try)):
            self.current_depth += 1
            self.max_depth = max(self.max_depth, self.current_depth)
        super().visit(node)
        if isinstance(node, (ast.If, ast.For, ast.While, ast.Try)):
            self.current_depth -= 1

    def analyze(self, code):
        try:
            tree = ast.parse(code)
            self.visit(tree)
            return {
                "identifier_length_violation": self.identifier_length_violation,
                "max_control_structure_nesting": self.max_depth,
            }
        except SyntaxError as e:
            return {"error": str(e)}

def analyze_file(filepath):
    with open(filepath, "r") as source_file:
        code = source_file.read()
    analyzer = ASTAnalyzer()
    result = analyzer.analyze(code)
    return result
